feat(runtime): support MIN_CORES, MAX_CORES, and MAX_CORES_RATIO
When the proxy boots up, it needs to select a number of I/O worker threads to
allocate to the runtime. This change adds a new environment variable that allows
this value to scale based on the number of CPUs available on on the host.

A MAX_CORES_RATIO value of 1.0 will allocate one worker thread per CPU core. A
lesser value will allocate fewer worker threads.

The MIN_CORES value sets a lower bound on the number of worker threads to use.
The MAX_CORES value sets an upper bound.