refactor(app/integration): call `tcp::client()` directly (#3694)
`linkerd_app_integration::tcp` provides a `TcpClient` type that is
distinct from the primary `linkerd_app_integration::client::Client` type
broadly used in integration tests.

this commit makes a small change to reduce indirection, and clarify that
this is constructing a different client implementation from a different
submodule.

this removes `linkerd_app_integration::client::tcp()`, and updates test
code to call the `tcp::client()` function that this is masking.

this is the client-side equivalent to #3688 (a10d1d7e).

Signed-off-by: katelyn martin <kate@buoyant.io>